Refactor comment display into a new element.
