+++ static/components.js
@@ -30,6 +30,7 @@ setBasePath('/static/shoelace');
 // chromedash components
 import './elements/icons';
 import './elements/chromedash-all-features-page';
+import './elements/chromedash-activity-log';
 import './elements/chromedash-app';
 import './elements/chromedash-approvals-dialog';
 import './elements/chromedash-banner';
